feat: datacollection dogtag cell type
