+++ packages/react/src/experimental/OneDataCollection/__stories__/cell-types.mdx
@@ -1,0 +1,279 @@
+import { Canvas, Meta, Controls, Unstyled } from "@storybook/blocks"
+import { DoDonts } from "@/lib/storybook-utils/do-donts"
+
+import * as DataCollectionStories from "./index.stories"
+import * as ActionDataCollectionStories from "./actions.stories"
+
+<Meta title="Data collection/Cell Types" />
+
+# Data collection Cell Types / Properties
+
+## Introduction
+
+In order to provide an homogeneus experience to the customers and abstract the
+render logic to developers, Data Collection's visualizacions ara limited to the
+following cell types
+
+Cell types also provide a semantic meaning to the data and allow visualizations
+to render the data according to the cell type, for example a `number` cell type
+renders the value as text and in the table visualizacion aligns it to right
+
+The way a cell/property must be rendered is defined in the `render` function of
+the datacollection, the type is the cell type to render and the value depends on
+that type
+
+## Cell types
+
+### text
+
+Renders a text This type's render function can be used as a shortcut returning
+just a string.
+
+```
+type value = string | undefined
+```
+
+Example:
+
+```
+render: (item) => 'my text'
+// or
+render: (item) => ({
+    type: 'text'
+    value: 'my text'
+})
+```
+
+### number
+
+Renders a number (right align on table visualizacion)
+
+#### Value type
+
+```
+type value = number | undefined
+```
+
+Example:
+
+```
+render: (item) => ({
+    type: 'number'
+    value: 1234
+})
+```
+
+### date
+
+Renders a date
+
+#### Value type
+
+```
+type value = Date | undefined
+```
+
+Example:
+
+```
+render: (item) => ({
+    type: 'date'
+    value: new Date()
+})
+```
+
+### amount
+
+Renders a monetary number (right align on table visualizacion)
+
+#### Value type
+
+```
+type value = number | undefined
+```
+
+Example:
+
+```
+render: (item) => ({
+    type: 'amount'
+    value: 1234
+})
+```
+
+### avatarList
+
+Renders a list of avatars
+
+#### Value type
+
+```
+type value = {
+    avatarList: AvatarVariant[]
+    max?: number // Max avatars to render
+}
+```
+
+Example:
+
+```
+render: (item) => ({
+    type: 'avatarList'
+    value: {
+        avatarList: [],
+        max: 3
+    }
+})
+```
+
+### status
+
+Renders a status tag
+
+#### Value type
+
+```
+type value = {
+    status: StatusVariant
+    label: string
+}
+```
+
+Example:
+
+```
+render: (item) => ({
+    type: 'status'
+    value: {
+        status: 'active',
+        label: item.name
+    }
+})
+```
+
+### person
+
+Renders a person avatar
+
+#### Value type
+
+```
+type value = {
+    firstName: string
+    lastName: string
+    src?: string // the avatar image
+}
+```
+
+Example:
+
+```
+render: (item) => ({
+    type: 'person'
+    value: {
+        firstName: 'Sergio'
+        lastName: 'Carracedo'
+        src: '/pth/to/the/avatar.png'
+    }
+})
+```
+
+### company
+
+Renders a company avatar
+
+#### Value type
+
+```
+type value = {
+    name: string
+    src?: string // the avatar image
+}
+```
+
+Example:
+
+```
+render: (item) => ({
+    type: 'company'
+    value: {
+        name: 'Factorial'
+        src: '/pth/to/the/avatar.png'
+    }
+})
+```
+
+### team
+
+Renders a team avatar
+
+#### Value type
+
+```
+type value = {
+    name: string
+    src?: string // the avatar image
+}
+```
+
+Example:
+
+```
+render: (item) => ({
+    type: 'team'
+    value: {
+        name: 'Factorial'
+        src: '/pth/to/the/avatar.png'
+    }
+})
+```
+
+### tag
+
+Renders a tag
+
+#### Value type
+
+```
+type value = {
+    label: string
+    icon?: IconType
+}
+```
+
+Example:
+
+```
+render: (item) => ({
+    type: 'tag'
+    value: {
+        name: 'Factorial'
+        src: FactorialLogoIcon
+    }
+})
+```
+
+### dotTag
+
+Renders a dot tag
+
+#### Value type
+
+```
+type value = {
+  label: string,
+  color: NewColor
+)
+```
+
+Example:
+
+```
+render: (item) => ({
+    type: 'dotTag'
+    value: {
+        label: 'Factorial'
+        color: 'malibu'
+    }
+})
+```
